added area to display calculations

+++ include/calc.h
@@ -5,12 +5,12 @@ class Calculator
 {
     public:
         Calculator();
-        double add(double x, double y);
-        double subtract(double x, double y);
-        double multiply(double x, double y);
-        double divide(double x, double y);
-        double raisePower(double x, double exponent);
-        double areaOfCircle(double radius);
+        long double add(long double x, long double y);
+        long double subtract(long double x, long double y);
+        long double multiply(long double x, long double y);
+        long double divide(long double x, long double y);
+        long double raisePower(long double x, long double exponent);
+        long double areaOfCircle(long double radius);
 };
 
 #endif
